rename png to image
